Renamed directory and files to better describe their purpose
